uugamebooster & luci-app-uugamebooster: (#6217)
* uugamebooster: rebuild startup service
* luci-app-uugamebooster: use standerd terms
* luci-app-uugamebooster: use inter-startup funcion
* luci-app-uugamebooster: fix typo
