+++ package/lean/luci-app-uugamebooster/Makefile
@@ -10,7 +10,7 @@ LUCI_DEPENDS:=+uugamebooster
 LUCI_PKGARCH:=all
 PKG_NAME:=luci-app-uugamebooster
 PKG_VERSION:=1.0
-PKG_RELEASE:=3
+PKG_RELEASE:=5
 
 include $(TOPDIR)/feeds/luci/luci.mk
