--- package/lean/luci-app-uugamebooster/luasrc/model/cbi/uuplugin/uuplugin.lua
@@ -1,15 +1,0 @@
-require("luci.util")
-
-mp = Map("uuplugin", translate("UU Game Accelerator"))
-mp.description = translate("A Paid Game Acceleration service")
-mp:section(SimpleSection).template  = "uuplugin/uuplugin_status"
-
-s = mp:section(TypedSection, "uuplugin")
-s.anonymous=true
-s.addremove=false
-
-o = s:option(Flag, "enabled", translate("Enable"))
-o.default = 0
-o.optional = false
-
-return mp
