+++ package/lean/luci-app-uugamebooster/luasrc/view/uugamebooster/uugamebooster_status.htm
@@ -1,0 +1,21 @@
+<script type="text/javascript">//<![CDATA[
+XHR.poll(3, '<%=url([[admin]], [[services]], [[uugamebooster]], [[status]])%>', null,
+	function(x, data) {
+		var tb = document.getElementById('uugamebooster_status');
+		if (data && tb) {
+			if (data.running) {
+				tb.innerHTML = '<em><b><font color=green>UU Game Booster <%:RUNNING%></font></b></em>';
+			} else {
+				tb.innerHTML = '<em><b><font color=red>UU Game Booster <%:NOT RUNNING%></font></b></em>';
+			}
+		}
+	}
+);
+//]]>
+</script>
+<style>.mar-10 {margin-left: 50px; margin-right: 10px;}</style>
+<fieldset class="cbi-section">
+	<p id="uugamebooster_status">
+		<em><%:Collecting data...%></em>
+	</p>
+</fieldset>
