--- package/lean/luci-app-uugamebooster/luasrc/view/uuplugin/uuplugin_status.htm
@@ -1,21 +1,0 @@
-<script type="text/javascript">//<![CDATA[
-XHR.poll(3, '<%=url([[admin]], [[services]], [[uuplugin]], [[status]])%>', null,
-	function(x, data) {
-		var tb = document.getElementById('uuplugin_status');
-		if (data && tb) {
-			if (data.running) {
-				tb.innerHTML = '<em><b><font color=green>UU GameAcc <%:RUNNING%></font></b></em>';
-			} else {
-				tb.innerHTML = '<em><b><font color=red>UU GameAcc <%:NOT RUNNING%></font></b></em>';
-			}
-		}
-	}
-);
-//]]>
-</script>
-<style>.mar-10 {margin-left: 50px; margin-right: 10px;}</style>
-<fieldset class="cbi-section">
-	<p id="uuplugin_status">
-		<em><%:Collecting data...%></em>
-	</p>
-</fieldset>
