+++ package/lean/luci-app-uugamebooster/po/zh-cn/uugamebooster.po
@@ -1,0 +1,9 @@
+
+msgid "Enable"
+msgstr "启用"
+
+msgid "UU Game Booster"
+msgstr "网易UU网游加速器"
+
+msgid "A Paid Game Acceleration service"
+msgstr "一个富家子弟用的游戏加速器"
