--- package/lean/luci-app-uugamebooster/po/zh-cn/uuplugin.po
@@ -1,12 +1,0 @@
-
-msgid "Enable"
-msgstr "启用"
-
-msgid "UU GameAcc"
-msgstr "UU游戏加速器"
-
-msgid "UU Game Accelerator"
-msgstr "UU游戏加速器"
-
-msgid "A Paid Game Acceleration service"
-msgstr "一个富家子弟用的的游戏加速器"
