+++ package/lean/luci-app-uugamebooster/root/etc/config/uugamebooster
@@ -1,0 +1,4 @@
+
+config uugamebooster 'uugamebooster'
+	option enabled '0'
+
