--- package/lean/luci-app-uugamebooster/root/etc/config/uuplugin
@@ -1,4 +1,0 @@
-
-config uuplugin 'uuplugin'
-	option enabled '0'
-
