--- package/lean/luci-app-uugamebooster/root/etc/init.d/uuplugin
@@ -1,23 +1,0 @@
-#!/bin/sh /etc/rc.common
-
-START=99
-STOP=10
-
-USE_PROCD=1
-
-NAME="uugamebooster"
-UU_BIN="/usr/share/uugamebooster/uuplugin"
-UU_CONF="/usr/share/uugamebooster/uu.conf"
-UU_LOGFILE="/tmp/uugamebooster.log"
-
-start_service() {
-  config_load uuplugin
-	local enable
-	config_get_bool enable uuplugin enabled
-	if [ "$enable" -eq 1 ]; then
-    procd_open_instance
-    procd_set_param command $UU_BIN $UU_CONF >$UU_LOGFILE 2>&1
-    procd_set_param respawn
-    procd_close_instance
-	fi
-}
