+++ package/lean/luci-app-uugamebooster/root/etc/uci-defaults/45_luci-uugamebooster
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@uugamebooster[-1]
+	add ucitrack uugamebooster
+	set ucitrack.@uugamebooster[-1].init=uugamebooster
+	commit ucitrack
+EOF
+
+rm -f /tmp/luci-indexcache
+exit 0
