--- package/lean/luci-app-uugamebooster/root/etc/uci-defaults/45_luci-uuplugin
@@ -1,11 +1,0 @@
-#!/bin/sh
-
-uci -q batch <<-EOF >/dev/null
-	delete ucitrack.@uuplugin[-1]
-	add ucitrack uuplugin
-	set ucitrack.@uuplugin[-1].init=uuplugin
-	commit ucitrack
-EOF
-
-rm -f /tmp/luci-indexcache
-exit 0
