+++ package/lean/uugamebooster/files/uugamebooster.init
@@ -1,0 +1,18 @@
+#!/bin/sh /etc/rc.common
+
+START=99
+STOP=10
+
+USE_PROCD=1
+
+NAME="uugamebooster"
+UU_BIN="/usr/share/uugamebooster/uuplugin"
+UU_CONF="/usr/share/uugamebooster/uu.conf"
+UU_LOGFILE="/tmp/uugamebooster.log"
+
+start_service() {
+	procd_open_instance
+	procd_set_param command $UU_BIN $UU_CONF >$UU_LOGFILE 2>&1
+	procd_set_param respawn
+	procd_close_instance
+}
